+++ Projects/Tools/ColladaToCSModel/src/com/chilliworks/chillisource/colladatocsmodel/ColladaToCSModelOptions.java
@@ -46,6 +46,8 @@ public final class ColladaToCSModelOptions
 	public boolean m_vertexHasPosition = false;
 	public boolean m_vertexHasTexCoords = false;
 	public boolean m_vertexHasNormal = false;
+	public boolean m_vertexHasTangent = false;
+	public boolean m_vertexHasBitangent = false;
 	public boolean m_vertexHasColour = false;
 	public boolean m_vertexHasWeights = false;
 	public boolean m_vertexHasJointIndices = false;